requirements.txt
# ==================================================================================
# Brain Connectivity-Based Cognitive Prediction - Python Dependencies
# ==================================================================================

# Core scientific computing
numpy>=1.21.0,<2.0.0
scipy>=1.7.0,<2.0.0
pandas>=1.3.0,<3.0.0

# Machine learning
scikit-learn>=1.0.0,<2.0.0

# Neuroimaging
nibabel>=3.2.0,<6.0.0
nilearn>=0.9.0,<1.0.0

# Visualization
matplotlib>=3.4.0,<4.0.0
seaborn>=0.11.0,<1.0.0
plotly>=5.0.0,<6.0.0

# Data handling
h5py>=3.0.0,<4.0.0
PyYAML>=5.4.0,<7.0.0

# Progress bars
tqdm>=4.60.0,<5.0.0

# Statistical testing
statsmodels>=0.13.0,<1.0.0
pingouin>=0.5.0,<1.0.0

# Utility
joblib>=1.0.0,<2.0.0
requests>=2.25.0,<3.0.0

# Report generation
jinja2>=3.0.0,<4.0.0

# Optional: Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=7.6.0

# Optional: Advanced visualization
ipython>=7.0.0
